Move TestHostBuilder into separate file
